RAG_Eval_Training/requirements.txt
# =============================================================================
# requirements.txt — Production-grade RAG Evaluation Training Project
# Install: pip install -r requirements.txt
# =============================================================================

# --- Core LLM / Embedding Clients ---
openai>=1.30.0
anthropic>=0.25.0

# --- Orchestration ---
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.2.0

# --- Vector Store (in-memory) ---
chromadb>=0.4.22

# --- Document Ingestion ---
pypdf>=4.0.0
python-dotenv>=1.0.0

# --- RAG Evaluation (RAGAS framework) ---
ragas>=0.1.14

# --- Utilities ---
pyyaml>=6.0
rich>=13.7.0          # Pretty terminal output for the report

# --- Web Framework ---
flask>=3.0.0
flask-cors>=4.0.0
